requirements.txt
accelerate==0.19.0
antlr4-python3-runtime==4.8
arrow==1.2.3
attrs==23.1.0
audiolm-pytorch==1.1.4
beartype==0.14.0
bitarray==2.7.4
boto3==1.26.151
botocore==1.29.151
bravado==11.0.3
bravado-core==5.17.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.35
einops==0.6.1
ema-pytorch==0.2.3
encodec==0.1.1
fairseq==0.12.2
filelock==3.12.0
fonttools==4.40.0
fqdn==1.5.1
fsspec==2023.5.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
grad-cam==1.4.8
huggingface-hub==0.15.1
hydra-core==1.0.7
idna==3.4
isoduration==20.11.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonref==1.1.0
jsonschema==4.17.3
kiwisolver==1.4.4
lion-pytorch==0.1.2
lit==16.0.5.post0
local-attention==1.8.6
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
musiclm-pytorch==0.2.2
neptune==1.2.0
networkx==3.1
nnAudio==0.3.2
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.0.6
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
portalocker==2.7.0
psutil==5.9.5
pycocotools==2.0.6
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
s3transfer==0.6.1
sacrebleu==2.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
swagger-spec-validator==3.0.3
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
triton==2.0.0
ttach==0.0.3
typing_extensions==4.6.3
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.16
vector-quantize-pytorch==1.6.11
wcwidth==0.2.6
webcolors==1.13
websocket-client==1.5.3
x-clip==0.12.1
